Refactor: modify update tags and title process
태그 테이블을 삭제하고 태그 이름의 배열을 저장하도록 수정하였습니다
