--- src/tag/dto/tag.dto.ts
@@ -1,20 +1,0 @@
-import { ApiProperty } from '@nestjs/swagger'
-
-class LinkDto {
-    @ApiProperty({ description: 'Link Id' })
-    linkId: number
-
-    @ApiProperty({ description: 'Link title' })
-    title: string
-
-    @ApiProperty({ description: 'tags' })
-    tags: string[]
-}
-
-export class CreateTagDto {
-    @ApiProperty({ description: 'User id' })
-    userId: number
-
-    @ApiProperty({ description: 'Links id with tags and title' })
-    links: LinkDto[]
-}
