--- src/tag/tag.controller.spec.ts
@@ -1,20 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing'
-import { TagController } from './tag.controller'
-import { TagService } from './tag.service'
-
-describe('TagController', () => {
-    let controller: TagController
-
-    beforeEach(async () => {
-        const module: TestingModule = await Test.createTestingModule({
-            controllers: [TagController],
-            providers: [TagService],
-        }).compile()
-
-        controller = module.get<TagController>(TagController)
-    })
-
-    it('should be defined', () => {
-        expect(controller).toBeDefined()
-    })
-})
