--- src/tag/tag.controller.ts
@@ -1,7 +1,0 @@
-import { Controller } from '@nestjs/common'
-import { TagService } from './tag.service'
-
-@Controller('tag')
-export class TagController {
-    constructor(private readonly tagService: TagService) {}
-}
