--- src/tag/tag.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common'
-import { TagService } from './tag.service'
-import { TagController } from './tag.controller'
-
-@Module({
-    controllers: [TagController],
-    providers: [TagService],
-})
-export class TagModule {}
