--- src/tag/tag.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing'
-import { TagService } from './tag.service'
-
-describe('TagService', () => {
-    let service: TagService
-
-    beforeEach(async () => {
-        const module: TestingModule = await Test.createTestingModule({
-            providers: [TagService],
-        }).compile()
-
-        service = module.get<TagService>(TagService)
-    })
-
-    it('should be defined', () => {
-        expect(service).toBeDefined()
-    })
-})
